+++ docs/conf.py
@@ -279,6 +279,7 @@ sphinx_gallery_conf = {
     # path to store the module using example template
     'mod_example_dir': 'modules/generated',
     'doc_module': ('km3pipe', ),
+    'default_thumb_file': '_static/default_gallery_thumbnail.png',
     'examples_dirs': ['../examples'],
     'gallery_dirs': ['auto_examples'],
 #     'reference_url':  {